facade_microservice/requirements.txt
fastapi
uvicorn
pyjwt
passlib[bcrypt]
pydantic
py-consul
httpx